refactor(maps): extract flask viewer + utilities + container detection
Split three self-contained clusters out of the 7000+ LOC MapsManager god
class so future refactors have smaller, focused targets:

- src/maps/_flask_viewer.py (1233 LOC): 10 module-level functions for
  the Flask-based interactive viewer, wired via callable injection
  (collect_payload_fn / build_response_fn) so the module has no
  dependency on MapsManager.
- src/maps/_maps_utils.py (133 LOC): flatten_dict_recursively,
  sanitize_filename, write_data_with_format_selection.
- src/maps/_container_detection.py (151 LOC): is_running_in_container
  with six independent heuristic checks.

Net: maps_manager.py -1361 LOC (7243 -> 5882). Compliance score
unchanged at F/54.0 due to the per-category 20pt severity cap; further
extractions (plotly, backup, wizard) needed to move the needle.

Checkpoint commit within the ongoing decomposition PR for rank 1 of
spec 1008 top-20 violations remediation.
